third_party/WebKit/LayoutTests/inspector/layers/layers-3d-view-hit-testing-expected.txt

Initial state
State of layers:
hovered layer: none
selected layer: none

Hovering content root
State of layers:
hovered layer: content root
selected layer: none

Selecting layer b
State of layers:
hovered layer: none
selected layer: layer b

Hovering layer a
State of layers:
hovered layer: layer a
selected layer: layer b

Selecting content root
State of layers:
hovered layer: none
selected layer: none
